requirements.txt
psutil
netifaces
djangorestframework
markdown
django-filter
requests